Too soon to rely on autosym on getLoadedAltAndSimplifiedName
Autosymbols are not yet ready be used to fetch simplified default names
when saving/restoring collection. Currently a module gets loaded with
the default auto-symbol even if there is an explicit default set for
this module in another enabled modulepath.

Also some environment gets loaded and then after user enables the
implicit_default and advanced_version_spec options, the autosymbols will
not be set but modulecmd will expect them when saving/restoring
collection.