Unset variant in collection means default value set
In a collection when a variant is not specified (no value set) on the
module load line it means this variant is set at the default value for
variant.

Code is adapted here in getVariantList and modEq to transmit the
is-default hint for each variant set on the loaded module against which
the module specification defined in collection is compared to.

The *missmeandfl* option is added to the modVariantCmp procedure to add
a specific comparison when enabled: if the variant set on loaded module
is not the default version but this variant is not set on the collection
definition, it means module specification in collection and loaded
module differ.