Update modnamevr after mod eval in execute-modulefile
Update the module name version with variant designation once modulefile
has been evaluated in the execute-modulefile procedure. After
evaluation, the variants with default value are set, so the designation
is then complete. Tags are then collected for this new designation has
it can change the tag list applying to loading module.

The up_namevr argument is added to the execute-modulefile procedure.
When set, it means the modnamevrvar argument is a variable name from the
calling context to bound to. Otherwise modnamevrvar contains the module
name version and variant string.

As the tags are fetched again after modulefile evaluation, the tags that
have been defined within modulefile are now collected and exported in
persistency variable.