Add variant support to stickyness handling
Adapt stickyness handling to support variant as, like any other tags,
stickyness may be applied to module specification including variants.

When testing if stickyness is reloading pass the currently module name
version with defined variants included.

Update isStickynessReloading procedure to handle module specification
with variants. Improve by the way the test to determine if tag applies
to module name and version and then if the expected designation is
reloading or not.