Preserve variant auto set dfl state when reloading
When running `reload` sub-command, ensure the automatically set default
value state of variant is preserved to end with the exact same loaded
module properties once reloading is finished.

getVariantList procedure is adapted to pass the '2' value to the
excl_dflval argument, which means that we want to exclude the default
value that have been automatically set (so when reloading they will be
this time also automatically set).