Add --boolean option to variant mfcmd
Add the --boolean option to the variant modulefile command to define a
boolean variant. When set, no list of accepted value should be defined
the variant and any default value set should be a boolean value.

When boolean variant is evaluated it is checked against any value
accepted as boolean by Tcl as described in Tcl manual:

"If string is any of 0, false, no, or off, then Tcl_GetBoolean stores a
zero value at *boolPtr. If string is any of 1, true, yes, or on, then 1
is stored at *boolPtr. Any of these values may be abbreviated, and
upper-case spellings are also acceptable."