Accept + char at end of module name
Accept any number of plus (+) character at the end of module name to be
part of this module name instead of being parsed as a boolean variant
character.

'+' character should not be followed by alpha character otherwise it
will be treated as part of boolean variant specification.

Thus 'name+' or 'name++' are accepted module names but 'name+name' are
taken as module name and enabled boolean variant.