Raise error when bool val set on non-bool variant
Raise an error when a non-boolean variant defines a boolean value in its
accepted value list. Integers which also are boolean (0 and 1) are yet
accepted on non-boolean variant.

As boolean value are translated in canonical form (0 or 1) when the
module specification is parsed, boolean value cannot be defined as
accepted value of non-boolean variant since it would never get matched.