Accept non-existent modulepath on use subcmd
Allow non-existent directory to be specified to use sub-command so
directory that does not exist yet could be added to MODULEPATH prior
being created.

Also it may allow a shared module configuration to define all existing
paths across the network but only some of these paths are enabled on a
given host.