install: set --with-initconf-in to etcdir by default
Installation option '--with-initconf-in' is set to 'etdir' by default
to locate configuration files in the directory designated by the
'--etcdir' option. Therefore the initialization configuration file is
named `initrc` in this directory, and the  modulepath-specific
configuration file is named `modulespath`.

Configuration files are thus located in a traditional etcdir rather
being part of the init dir among scripts that should stay read only.