ts: introduce quick test mode
Introduce the non-regression quick test mode. When the `QUICKTEST`
environment variable is set to 1, only the main tests from the
non-regression testsuite are run.

When first argument of the `mt` script is `quick`, tests are run in
quick mode.

Goal is to complete the quick tests in a minute, rather the 10+ minutes
required by the full testsuite.