No ref count for Modules-specific path variables
Exclude the Modules-specific variables (like LOADEDMODULES) from the
reference counting mechanism. As a result no <VAR>_modshare variable is
set in user environment for these variables. Exception is made for
MODULEPATH environment variable that still benefit from the mechanism.

This change is made as the reference counting mechanism is useless for
all Modules-specific variables except for MODULEPATH. Either the entries
are unique (same module cannot be loaded twice in LOADEDMODULES) or
duplicate elements are allowed (same modulefile could be used twice by
virtual modules in _LMFILES_). Also each entry in the __MODULES_LM*
variables uses as prefix the loaded module name the entry is associated
to.