Use '__MODULES_SHARE_' prefix on ref count env var
Rename the environment variables used to record the reference count of
each element in path-like environment variables. A `__MODULES_SHARE_`
prefix is applied to the name of these variables instead of the
`_modshare` suffix to indicate they are intended for Modules internal
use of only.

It also helps with DYLD_* environment variables where using a
suffixed-variable led to warning messages on OS X.