Update ref count var when element is added several times
When an element is added to a path-like variable through the
append-path or prepend-path modulefile commands, add this element to the
associated reference counter variable only when this element is added
multiple times to the path-like variable. When an element is removed
from a path-like variable, this element is removed from the reference
counter variable when its counter is equal to 1.

Exception is made when element is an empty string: in this case element
is added to the reference counter variable even if only added once to
the path-like variable in order to distinguish between an empty path-like
variable and a path-like variable containing an empty string as single
element.