Add isTopEvaluation proc
Gather in a dedicated procedure named isTopEvaluation the test to know
if current evaluation is made at top level.

"Top level" is either: (1) at the modulecmd level (module command
written by user in terminal or script) or (2) during  the evaluation of
rc or modulefile by a source or autoinit sub-command triggered from
modulecmd level or (3) during the evaluation of global rc file.

(2) and (3) are considered "extended" top evaluation contexts.