Add --ignore-refcount opt to add-path/unload-path proc
Add the '--ignore-refcount' option to the add-path and unload-path
procedures (parsed by parsePathCommandArgs procedure). When set this
option makes the reference counter value ignored.

Which means on add-path that the reference counter of each path element
passed is not increased unless if the duplication mode is enabled.

On unload-path it means that the reference counter is not take into
account and each path element is removed (even if the reference counter
is greater than 1).