Revive the refresh subcmd
Restore the refresh sub-command as it worked in version 3.2: not
anymore an alias on the reload sub-command but an evaluation of each
loaded module (in their loaded order) in 'refresh' mode.

`refresh` evaluation mode only execute the modulefile commands
generating volatile environment changes like set-alias and set-function.
variant and source-sh commands should also be enabled to ensure a
modulefile evaluation that regenerates the expected volatile environment
changes.

`refresh` evaluation mode does not require that the modulefile commands
triggering environment variable changes (setenv, unsetenv, append-path,
etc) to operate as the targetted variables should already be set as the
refreshed modules are loaded. Same goes for the x-resource and system
modulefile commands. Configuring these commands as 'no operation' for
this refresh mode should not disrupt the modulefile evaluation.

refresh sub-command does not accept any argument and cannot be called
within modulefile or sourced or initilization module script (initrc).

When the evaluation of a refreshed modulefile fails, the environment
changes prior this evaluation is restored. Then refresh evaluation
continue on the remaining loaded modules.