Refresh loaded modules when initializing
When initializing Modules, refresh the loaded modules in case some user
environment is already configured. This is especially useful when
starting a sub-shell session as it ensures that the loaded environment
in parent shell will be correctly inherited, as the `refresh`
sub-command re-applies the non-persistent environment configuration
(i.e., shell alias and function that are not exported to the sub-shell).

If a refresh has to be attempted, loaded environment is parsed. In case
this environment is inconsistent, an error code is returned but this
error is trapped and initialization evaluation continues without
launching the refresh process. No error message is reported to avoid
noise on the full shell session initialization. The user will get the
error message as soon as he will launch the next module command that
parse the loaded environment.

Fixes #86.