Eval both modulespath and initrc when initializing
When initializing Modules, evaluate the initrc configuration file in
addition to the the modulespath configuration file and not instead of
this file. initrc is evaluated after modulespath file.

Both configuration files are only evaluated if the current environment
if found unset, in other words when the MODULEPATH and LOADEDMODULES
environment variables are both found unset or empty.

As initrc is evaluated after modulespath, it has the ability to undo
what was done by modulespath, especially if it restores a module
collection.