ts: fix install tests when dfl loadedmodules are set
When the loadedmodules installation option is set and if the module
initialization is triggered then the default module to load will emit a
message on stderr. Adapt the testsuite code, especially the _test_sub
procedure to automatically take this situation into account (no need to
set these load message in each expected stderr answer).