Set tag on modvr desig of loaded mod
Correctly detect tags set on loaded modules when refreshing them.

Update the cacheCurrentModules procedure to record the tags of loaded
modules also on the module name version and variant list designation. So
when the `module-info tags` command is called from a modulefile loaded
with variant set, it correctly retrieves the tags defined on this module
if currently loaded.

Add the report mode '6' on getVariantList procedure to proceed like mode
'1' but without updating the key map reference.