Ensure no autoinit loop when set_shell_startup set
Set the __MODULES_AUTOINIT_INPROGRESS environment variable when running
the `autoinit` sub-command and quit autoinit process if this variable is
found defined and equal to 1 when starting it. Ensure this way that an
autoinit process will not be triggered indefinitely by itself when the
set_shell_startup option is enabled and some module loaded at
initialization time relies on the execution of a bash script.

Fixes #414.