Add --redirect/--no-redirect cmdline switches
Add the --redirect and --no-redirect command line switches to control
where the output of the module command should be sent. When redirected
it is sent to stdout otherwise not redirected means it is left on
stderr.

These command line switches have no effect within modulecmd.tcl code.
They are treated directly in module shell function to adapt output
redirection.

Closes #410.