Add output redirection handling on sh module function
Update the definition of the module function for sh shell family (sh,
bash, ksh and zsh) to support the output redirection.

Now both _module_raw and module shell functions are always defined.
Default redirection behavior depends on the interactiveness of the shell
when the autoinit command is run. Then shell code is set in module
function definition to adapt behavior depending on
MODULES_REDIRECT_OUTPUT environment variable value or
--redirect/--no-redirect command-line switches.