Reuse more path list defined for load in parsePathCommandArgs
In case the --append-on-unload argument is set after environment
variable name but prior path list definition, the path list to use for
load evaluation will also be used on the append behavior during unload
evaluation.

In addition, if no path list is defined after --append-on-unload
argument, the list used for load evaluation is reused on the append
behavior of unload evaluation.

Applies to remove-path and module unuse modulefile commands.