Add --prepend-on-unload on remove-path/module unuse
Add the --prepend-on-unload option to the remove-path and module unuse
sub-command to append specified paths when commands are evaluated in
unload mode.

If option is set prior path list specification, this path list which is
unsets when modulefile loads is appended back when modulefile unloads.
If a different path list is passed after option specification, this list
is appended on unload mode instead of the list removed at load time.

--prepend-on-unload option cannot be mixed with --index option.