Pass and apply tag_list in cmdModuleLoad
Transmit the tag list specified on module load, try-load or switch to
the cmdModuleLoad procedure to apply these tags to the module to load.

Need to refine later on tag list transmission on collection management
(cmdModuleRestore) and environment reload (reloadModuleListLoadPhase).

An already loaded module will not be reloaded if a different tag list is
specified. Loaded tags for this module are not changed either.