Save tags set on loaded modules in collection
Adapt collection mechanism to dump the tags set on loaded module when
saving a collection and setting them back when collection is restored.

Tags are recorded in collection with the --tag= option set on module
load lines. Previously used --notuasked option is still processed if
found in collection but when saving new ones the auto-loaded tag is
dumped instead 'module load --tag=auto-loaded foo/1.0' for instance.

With this current patch savec tags are restored if module is not yet
loaded at the correct position in expected loaded module list.