Introduce load-any sub-command
Add the load-any module sub-command that loads first valid module in
specified list. Errors for unknown modulefile are silenced like for
try-load sub-command. add-any is an alias over load-any sub-command.

The any_modulefile state is introduced to indicate cmdModuleLoad to stop
after first loaded modulefile in list.

When used from a modulefile evaluation context, load-any first checks if
a module from its specified list is already loaded. If this is the case,
no additional requirement load is performed as constraint is considered
satisfied.