No mcookie check when searching module
Update checkValidModule to skip Module magic cookie check in file if
mcookie_check configuration option is not set to 'always'.

When mcookie_check is set to 'eval', files are not read when searching
modulefiles. Which means any file located within a modulepath directory
is considered a modulefile. An error will be obtained if an attempt is
made to evaluate a non-modulefile located in a modulepath directory.

Also if a modulefile is specified fullpath, Modules magic cookie will
also not be checked if mcookie_check is set to 'eval', and if this file
is not a modulefile an error will be obtained if an attempt is made to
evaluate this file.