Add family modulefile command
Introduce the family modulefile command that indicates currently loading
module provides name passed as argument and this name could only be
provided by one loaded module.

family is implemented with the combined definition of a conflict and a
module alias.

An environment variable is set to indicate what module currently
provides for the family within the loaded environment.