Add search capabilities to list subcmd
When module specification is passed to the list sub-command, only output
the loaded modules that matches at least one of these specification (OR
operation).

Sub-command header message is adapted when module specifications are
passed to the sub-command. However same header message is reported if
there is no loaded module to report from the start.

Search capabilities rely on the modEq comparison procedure, so it takes
into account the search_match mode (starts-with/contains), the
alternative names set over loaded modules and the variant defined.

Only the first specified search pattern is set for SGR highlight due to
the current limitation of reportModules procedure (which only accepts
one mod argument).