Add complete/uncomplete modulefile commands
Introduce the complete and uncomplete modulefile commands to define
completion mechanism on running shell. It supports only shell with
completion capabilities that could be defined with shell commands: bash,
tcsh and fish.

complete command takes a shell name as first argument to apply
completion specification only if currently running shell corresponds to
this shell name set as argument. This is done as completion
implementation is really different from one shell to another. Also the
arguments passed after command name are specific of the shell and are
not checked by module command.

When unloading module, complete becomes uncomplete. uncomplete can also
be set as an independent modulefile command and only acts on load
evaluation module. complete is also evaluated in refresh evaluation mode
to push completion definition in sub-shells for instance.

Multiple completion definition for the same command may be defined and
will be applied to shell. Some shell, like fish, needs multiple complete
commands to fully define all parameters for the completion of a single
command.