Enclose modulecmd.tcl output in quotes on sh-kind shells
Update the module shell function definition for sh-kind shell to enclose
the output generated by modulecmd.tcl in quotes in order to pass it to
the eval command.

Quoting code to pass it to eval brings correctness to the definition of
shell functions without harming already well functionning definition.
Regular shell function definition could now be correctly defined in user
environment: no more need to get each command line in these function
ending with a ';' character.

Same mechanism was applied in initialization script to evaluate the code
produced by the autoinit sub-command.

Multiple function-definition issues spotted when using source-sh across
environment setting script shell are fixed with this change.

Moreover, all the noglob_eval_shells work is made useless with this
change as no globbing is applyed to the code produced by modulecmd.tcl
as it is now enclosed in quotes.