Add pushenv modulefile command
Introduce the pushenv modulefile command that sets an environment
variable with a value specified as argument but saves the previous value
set to restore it when modulefile is unloaded.

Use a __MODULES_PUSHENV_<VAR> environment variable as a stack to record
the previous values of environment variable <VAR>. Each element in this
Modules-specific variable is the combination of the currently evaluating
modulename and pushed value. When unloading, the value set by this
module is removed not the value on top of the list.

When saving value set before any module, an empty module name is used
to push to the stack. When restoring this initial value, initial entry
in stack is also cleared (as no other module unload will unset it).

Follow same implementation principle that setenv, with a dedicated
procedure for the unload and whatis modes.