Add Module header when saving coll with tags
Record Modules magic cookie at the beginning of a collection when tags
are found set on the module to save. The #%Module file signature is
followed by the version number of Modules introducing the tag support.

This version number is temporarily set to 5.0 for test purpose. It will
be moved to 5.1, on the release commit of 5.1.0.

If collection starts with a Modules file signature, this header will not
be reported on saveshow output.