Check collections are valid when searching them
Update findCollections procedure to rely on a checkValidColl procedure
that verifies given collections can be read and are compatible with
current Modules version if they express a minimum version requirement.

As a side effect, when listing available collections with savelist
sub-command, now the collections that cannot be read are withdrawn from
result list.