Quote name if it contains space when listing modules
Update formatListEltToDisplay and formatListEltToLongDisplay to enclose
element name in single quote if it contains a space character. As a
result list and avail output will now report enclosed module names if
space character if found in them, like it is done when reporting module
designation as loading/unloading header.

Only enclose element name and not suffix, to follow the same output
approach than 'ls' command.