Create auto_handling and not version of prereq procs
Split prereq and prereq-all procedures into a auto_handling version,
that attempts to automatically load requirements, and a non
auto_handling version that raise an error directly if requirement is not
already loaded.

Adapt Tcl sub-interpreter initialization to make it dependent of the
auto_handling option enablement. When auto_handling is enabled, the
auto_handling versions of the prereqs procedure are configured.

With this change other parts of the code will be able to call either
prereq or prereqAH directly whether they need a load attempt or not,
whatever the current status of the auto_handling option.