Add --optional option to prereq commands
Introduce the --optional option on prereq commands (prereq, prereq-all,
always-load, prereq-any and depends-on) to indicate that the prereq list
is fully optional: if none of these prereqs are loaded, no error occurs.

When recording these prereq list in persistency information
(__MODULES_LMPREREQ variable), the current module name is added as first
module in prereq list to get this rule always satisfied even if no
module from the prereq list is loaded.