Use optional prereq to record try-load requirements
When try-load is used in a modulefile, record implied requirement as
optional instead of only recording those that are loaded. This way the
modulefile will get reloaded if requirement is loaded afterward.

Cannot fully rely on prereqAllModfileCmd procedure as try-load does not
fully correspond to a prereq-all --optional: an requirement evaluation
error will lead to an error of the module try-load command that will
abort the modulefile evaluation.