Export module bash functions directly through autoinit
Export the module, _module_raw and ml bash shell functions directly
through the autoinit sub-command process rather within the bash shell
initialization script.

Cannot apply the same change to the sh shell when it is bash launched as
sh, since we cannot tell in the modulecmd.tcl execution that the current
sh shell is bash (as BASH environment variable is not exported to
sub-shells).

Fixes #462.