Update autoinit produced code to enclose install path
Update code generated by the autoinit sub-command to enclose all
references to installation path to cope with an installation done in a
path containing whitespace character.

Enclosing modulecmd.tcl installation path this way is not effective with
the quarantine mechanism that implies a double shell code evaluation. So
it is advised not to install Modules in a path containing a whitespace
character if the quarantine mechanism is needed.