cirrus: test install path containing space char
Transform the second FreeBSD test case into a Ubuntu testcase, running
on last LTS release. For this test case, installation is made over a
path containing a whitespace character.

No specific quarantine setup with this test case, as it is not
compatible with installation in path containing a whitespace character.