Output produced shell code in one puts call
Instead of outputting shell code to sdtout for each line, collect all
produced shell code in a list variable (g_shcode_out) and output all
the code in a single puts call at the end of module command execution.

Create a specific renderFlush procedure to make the puts to stdout as
output could occur after renderSettings but also after
reportErrorAndExit and the exit catch branch in the main procedure.
renderFlush is called in cleanupAndExit procedure, as it is always run
in the end of the execution whatever if everything goes well or if error
occurs.

Code of the putsModfileCmd procedure is also adapted to build list of
message to output to stdout with a flag to know if end-of-line should be
added at end of message. Previously this procedure was listing arguments
to directly pass to puts calls.