Add search capabilities to savelist subcmd
When one or multiple strings are passed to the savelist sub-command,
only output the collections that match at least one of this string (OR
operation).

Sub-command header message is adapted when strings are passed to the
sub-command. However same header message is reported if there is no
saved collection to report from the start.

Search capabilities rely on the modEq comparison procedure, so it takes
into account the search_match mode (starts-with/contains) and the
insensitive case search mode.

Icase mode is set to search by default for savelist sub-command.