Allow unload of sticky modules on restore sub-cmd
Allow the unload of sticky modules when restoring collection. Goal is to
be able to restore designated collection as it was saved, thus without
the loaded sticky modules. If such collections were saved without the
sticky modules in it, it means that these sticky modules can safely be
unloaded when moving to the collection state.

As a result, using the --force option is of no use for restore
sub-command.

This change does not impact super-sticky modules. Such modules cannot be
unloaded during a restore, even if forced.

With this change, behavior of restore sub-command with sticky modules is
aligned with Lmod.