lint(ShellCheck): fix SC2128 warning
These occurrences do not cause any problems described below but it's better
to be verbose that we indeed want to access the first array element.

According to ShellCheck's documentation:

When referencing arrays, $myarray is equivalent to ${myarray[0]} --
it results in only the first of multiple elements.

Related: https://www.shellcheck.net/wiki/SC2128
Related: #470