Add 'stashpop' sub-command
Introduce stashpop sub-command to restore a stash collection then delete
it.

This sub-command cannot be called from a modulefile evaluation context.

Add sub-command empty skeleton: cmdModuleStashpop.

Sub-command accepts no option. It may take an optional argument: the
index of the stash collection to restore or the name of the stash
collection. If no argument is provided, stash collection 0 is assumed.