Add no_other_target option to findCollections
Add no_other_target option to findCollections procedure to filter when
glob searching the result from other target. Especially useful when
collections for specific targets exist but no target is currently set in
user environment.

This no_other_target is disabled by default. savelist sub-command
currently returns other targets when no target is set. But the option
will be enabled for stashpop, to ensure that the popped collection is
from the current target or from the *no target* if none set.