Revert "init: setup zsh FPATH through autoinit command"
This reverts commit 59ac24e2431df012d8b22872a0e8fd1e86a20e9f.

FPATH variable cannot be handled within autoinit process as this is not
an environment variable. Thus when modulecmd.tcl is executed, it does
not know the current value of FPATH. So if it sets FPATH during
autoinit, previous value of this variable is lost. As a result FPATH
must stay defined in initialization script.