Core code for cmdModuleCachebuild
Implement cachebuild sub-command. For each modulepath enabled or passed
as argument, if modulepath is writable for user build cache content for
this modulepath and record it in cache file.

Report message block is used to report file creation. Reports are output
in normal verbosity level or higher. A warning message is emitted if
modulepath is not writable for user, or if cache content to create is
empty.

Provide empty skeleton for formatModuleCacheContent procedure.