Introduce cache_buffer_bytes configuration option
Add cache_buffer_bytes configuration option to control the buffer size
used when reading or writing cache files. A 32 kbytes buffer is set by
default. It can vary from 4 kbytes to 1 MB. It is useful to have a large
buffer when reading or writing large files to reduce the number or read
or write system calls.

When configuration option is set, it defies the
MODULES_CACHE_BUFFER_BYTES environment variable.