Disable auto symbols on wild search with variant
Do not compute auto symbols if search query does not contain a module
name and version but contain variant specification.

Add isSpecWildWithVariant procedure to determine if search query is wild
with variant specification.

Disabling auto symbols is especially useful for sub-commands like
'paths' that take getModules result and try to resolve each entry. In
this situation it is important not to have any auto symbol that
indirectly targets an element withdrawn from extra match filter.