Raise error when extra spec on non-search cmds
Extra specification in module search query is only allowed on search
sub-commands (avail, paths, whatis). An error is raised when parsing a
module specification if it contains an extra specification on a
non-search context.

xtspec argument is added to parseModuleSpecification procedure to know
if extra specification is allowed or not.