Parse mod spec in extra specifier value with getavails on
For some extra specifier, a module specification is expected as value.
For them parse this module specification with resolution of all matching
available modules.

During extra match search process, test match of specification against
all matching available modules recorded when parsing spec.

modEq procedures have been updated to fetch available module names
recorded in module version spec structure as alternative names for spec
when ismodlo argument is set to 5.

modVariantCmp procedure has been updated to let a missing variant
definition of mod be considered as a match. As compared pattern and mod
are switched on extra specifier context (pattern is definition found in
modulefile and mod is also a pattern used for extra specifier definition
on command line).

Note that this methodology is not yet perfect as aliases, symbolic
versions and modules that do not exist are not part of the list of
available modules returned.